backend/query-visualization/requirements.txt
oracledb
fastapi
uvicorn
pandas
plotly
seaborn
python-dotenv
pytest
openai
pymongo
